refactor: move func CollectRuntimeInfosFromPVCs to webhook utils package
